--- a/lib/ar2dto/dto.rb
+++ b/lib/ar2dto/dto.rb
@@ -6,17 +6,22 @@ module AR2DTO
       Class.new(self) do
         include ::AR2DTO::ActiveModel if original_model.ar2dto.active_model_compliance
         define_singleton_method(:original_model) { original_model }
+        attr_reader(*original_model.attribute_names - original_model.ar2dto.except)
+
+        private
+
+        attr_writer(*original_model.attribute_names)
       end
     end
 
     def initialize(data = {})
-      @data = data.symbolize_keys
+      attribute_names.each { |attribute| send("#{attribute}=", data[attribute]) }
+      @data = data.except(*attribute_names).symbolize_keys
       super()
     end
 
     def as_json(options = nil)
-      attribute_names = self.class.original_model.attribute_names.map(&:to_sym)
-      @data.slice(*attribute_names).as_json(options)
+      super(options).except("data")
     end
 
     def ==(other)
@@ -27,12 +32,18 @@ module AR2DTO
       end
     end
 
+    private
+
+    def attribute_names
+      self.class.original_model.attribute_names
+    end
+
     def respond_to_missing?(name, include_private = false)
-      @data.key?(name) || super
+      @data&.key?(name) || super
     end
 
     def method_missing(method, *args, &block)
-      return @data[method] if @data.key?(method)
+      return @data[method] if @data&.key?(method)
 
       super
     end