--- a/spec/ar2dto/dto_spec.rb
+++ b/spec/ar2dto/dto_spec.rb
@@ -29,7 +29,7 @@ RSpec.describe AR2DTO::DTO do
 
     it "implements equality" do
       expect(subject).to eq(klass.new(user.attributes))
-      expect(subject).to_not eq(klass.new(user.attributes.merge(first_name: "Other")))
+      expect(subject).to_not eq(klass.new(user.attributes.merge("first_name" => "Other")))
     end
 
     it "does not respond to other methods" do
@@ -49,7 +49,7 @@ RSpec.describe AR2DTO::DTO do
     end
 
     context "with options" do
-      let(:options) { { except: :first_name } }
+      let(:options) { { except: "first_name" } }
 
       it "affects the result" do
         expect(subject).to eq(user.attributes.as_json.except("first_name"))