EDIT: still experimenting with changes

This solution is uglier but means a 8x improvement in performance, so I think it's worth it 😅 

```
Warming up --------------------------------------
       normal method   415.535k i/100ms
    singleton method     9.495k i/100ms
      method missing    85.462k i/100ms
Calculating -------------------------------------
       normal method      4.168M (± 1.1%) i/s -     21.192M in   5.084880s
    singleton method    117.048k (±12.7%) i/s -    579.195k in   5.015340s
      method missing    861.084k (± 2.2%) i/s -      4.359M in   5.064503s

Comparison:
       normal method:  4168169.3 i/s
      method missing:   861084.3 i/s - 4.84x  (± 0.00) slower
    singleton method:   117047.7 i/s - 35.61x  (± 0.00) slower
```

results for 

```ruby
# frozen_string_literal: true

require "bundler/inline"

gemfile(true) do
  source "https://rubygems.org"

  git_source(:github) { |repo| "https://github.com/#{repo}.git" }

  gem "benchmark-ips"
end

class Normal
  def e_0; end
  def e_1; end
  def e_2; end
  def e_3; end
  def e_4; end
end

class SingletonMethod
  def initialize(hash = {})
    hash.each { |key, value| define_singleton_method(key) { value } }
    super()
  end
end

class MethodMissing
  def initialize(hash = {})
    @hash = hash
  end

  def respond_to_missing?(name, include_private = false)
    @hash.has_key?(name) || super
  end

  def method_missing(method, *args, &block)
    return @hash[method] if @hash.has_key?(method)

    super
  end
end

Benchmark.ips do |x|
  x.report("normal method") do
    en = Normal.new
    en.e_0
    en.e_1
    en.e_2
    en.e_3
    en.e_4
  end

  x.report("singleton method") do
    e = SingletonMethod.new(e_0: nil, e_1: nil, e_2: nil, e_3: nil, e_4: nil)
    e.e_0
    e.e_1
    e.e_2
    e.e_3
    e.e_4
  end

  x.report("method missing") do
    e = MethodMissing.new(e_0: nil, e_1: nil, e_2: nil, e_3: nil, e_4: nil)
    e.e_0
    e.e_1
    e.e_2
    e.e_3
    e.e_4
  end

  x.compare!
end
```